feat(spm): root Package.swift with useLocalNatives toggle, fix MPS link
Move the Swift Package manifest to the repo root so SPM can resolve it via
the public GitHub URL with a stable tag. The previous setup pointed users at
either the `swift` branch (manifest missing MetalPerformanceShaders/MPSGraph
linkers — `_MPSSupportsMTLDevice` undefined at link time) or `from: "0.1.0"`
(no stable tag exists yet, and SPM `from:` excludes pre-release identifiers).

The new root manifest:
- links Metal, MetalPerformanceShaders, MetalPerformanceShadersGraph, CoreML,
  Accelerate, Security, c++ — fixes the actual MPS link bug
- gates XybridFFI between a local xcframework path and a versioned GitHub
  release-asset URL via a `useLocalNatives` Bool, mirroring the
  RunanywhereAI/runanywhere-sdks pattern
- ships scripts/set-natives-mode.sh (--set-local | --set-remote | --status)
  and scripts/sync-spm-checksum.sh (--check for CI, patch-in-place for
  release prep) under bindings/apple/scripts/

Release workflow:
- Adds a fail-fast verify step that runs sync-spm-checksum.sh --check against
  the freshly-built XybridFFI zip, so a tag whose root Package.swift is
  stale fails CI before any release asset is published.
- Backfills MetalPerformanceShaders + MPSGraph linker frameworks into the
  legacy `swift` branch publish-swift job so existing pinned consumers
  aren't broken while the branch still exists.
- Advertises both the new `exact:` install snippet and the deprecated
  `branch: "swift"` snippet in release notes.

Docs/READMEs (root + ja + zh, quickstart and SDK pages):
- Replace `from: "0.1.0"` and `branch: "swift"` with
  `exact: "0.1.0-beta13"` and a note that `from:` will work once a stable
  0.1.0 ships.

Verified locally:
- `swift build` against root Package.swift with useLocalNatives = true
  (xcframework built via `cargo xtask build-xcframework`) succeeds.
- A separate consumer package depending on this repo via `path:` builds and
  runs against XybridModelLoader — proves the manifest plus binary plumbing
  resolve cleanly via SPM.

Tagging v0.1.0-beta13, pushing the tag, and deleting the legacy `swift`
branch are deliberately left as follow-ups for a maintainer to authorize.
