--- bindings/apple/Package.swift
@@ -1,42 +1,0 @@
-// swift-tools-version:5.7
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
-import PackageDescription
-
-let package = Package(
-    name: "Xybrid",
-    platforms: [
-        .iOS(.v13),
-        .macOS(.v10_15)
-    ],
-    products: [
-        .library(
-            name: "Xybrid",
-            targets: ["Xybrid"]
-        )
-    ],
-    targets: [
-        // Main Swift target with public API
-        .target(
-            name: "Xybrid",
-            dependencies: ["XybridFFI"],
-            path: "Sources/Xybrid",
-            linkerSettings: [
-                .linkedLibrary("c++"),
-                .linkedFramework("Metal"),
-                .linkedFramework("MetalPerformanceShaders"),
-                .linkedFramework("MetalPerformanceShadersGraph"),
-                .linkedFramework("CoreML"),
-                .linkedFramework("Accelerate"),
-                .linkedFramework("Security"),
-            ]
-        ),
-        // Pre-built XCFramework containing the Rust FFI static library + headers
-        // For local development: build with `cargo xtask build-xcframework`
-        // For distribution: replaced with URL-based binaryTarget in release tags
-        .binaryTarget(
-            name: "XybridFFI",
-            path: "XCFrameworks/XybridFFI.xcframework"
-        )
-    ]
-)
