+++ docs/content/docs/quickstart.zh.mdx
@@ -52,9 +52,11 @@ https://github.com/xybrid-ai/xybrid
 
 ```swift
 dependencies: [
-    .package(url: "https://github.com/xybrid-ai/xybrid", from: "0.1.0")
+    .package(url: "https://github.com/xybrid-ai/xybrid", exact: "0.1.0-beta13")
 ]
 ```
+
+> beta 阶段请用 `exact:`，稳定版 `0.1.0` 发布后即可改用 `from: "0.1.0"`。
 
 **运行推理**
